requirements.txt
arrow==1.2.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
cookiecutter==1.7.3
Django==3.2
django-browser-reload==1.3.0
django-ckeditor==6.2.0
django-environ==0.8.1
django-js-asset==1.2.2
django-parler==2.3
django-redis-cache==2.1.0
django-rq==2.1.0
django-tailwind==3.1.1
django-translation-manager==1.1.1
djangorestframework==3.13.1
idna==3.3
Jinja2==3.0.3
jinja2-time==0.2.0
MarkupSafe==2.0.1
mysqlclient==2.1.0
polib==1.1.1
poyo==0.5.0
python-dateutil==2.8.2
python-slugify==5.0.2
pytz==2021.3
redis==3.5.3
requests==2.27.1
rq==1.10.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
whitenoise==5.3.0
